requirements.txt
torch >= 1.3.0
torchvision
advertorch
tqdm
pillow
imagenet_c